requirements.txt
streamlit==1.32.2
streamlit_extras==0.4.0
streamlit_js_eval==0.1.7
extra_streamlit_components==0.1.70
langchain==0.1.12
langchain_community==0.0.28
langchain_core==0.1.32
langchain_openai==0.0.8
openpyxl==3.1.2
Pillow==10.2.0
python-dotenv==1.0.1
watchdog==4.0.0